work on the OSI pluginSDK, added a bunch of new functionality
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@10554 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ CPRVolumeData.h
@@ -202,8 +202,8 @@ v   =  t0 + (z)*(t1-t0);
         int xinc, yinc, zinc;
         
         xinc = 1;
-        yinc = inlineBuffer->pixelsWide;
-        zinc = inlineBuffer->pixelsWideTimesPixelsHigh;
+        yinc = (int)inlineBuffer->pixelsWide;
+        zinc = (int)inlineBuffer->pixelsWideTimesPixelsHigh;
         
         // Compute the increments to get to the other 7 voxel vertices from A
         Binc = xinc;
